# resources/openapi-mt.json
{
  "components": {
    "schemas": {
...
      "UploadContext": {
        "properties": {
          "fields": {
            "items": {
              "items": false,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "type": "array"
          },
          "headers": {
            "items": {
              "items": false,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "type": "array"
          },
          "method": {
            "type": "string"
          },
          "uploadToken": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string"
          },
          "useMultipart": {
            "type": "boolean"
          }
        },
        "required": [
          "uploadUrl",
          "method",
          "useMultipart",
          "headers",
          "fields",
          "uploadToken"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "bearerFormat": "AccessToken",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "dev@kamu.dev",
      "name": "Kamu Data Inc."
    },
    "description": "Decentralized data management tool",
    "license": {
      "name": ""
    },
    "title": "kamu-cli",
    "version": "0.205.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/odata/{account_name}": {
      "get": {
        "operationId": "odata_service_handler_mt",
        "parameters": [
          {
            "description": "Account name",
            "in": "path",
            "name": "account_name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "OData root service description",
        "tags": [
          "kamu-odata"
        ]
      }
    },
...
    "/odata/{account_name}/{dataset_name}": {
      "get": {
        "operationId": "odata_collection_handler_mt",
        "parameters": [
          {
            "description": "Account name",
            "in": "path",
            "name": "account_name",
            "required": true
          },
          {
            "description": "Dataset name",
            "in": "path",
            "name": "dataset_name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "OData collection",
        "tags": [
          "kamu-odata"
        ]
      }
    },
    "/platform/file/upload/prepare": {
      "post": {
        "operationId": "platform_file_upload_prepare_post_handler",
        "parameters": [
          {
            "in": "path",
            "name": "fileName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contentLength",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "contentType",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadContext"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "summary": "Prepare file upload",
        "tags": [
          "kamu"
        ]
      }
    },
    "/platform/file/upload/{upload_token}": {
      "get": {
        "operationId": "platform_file_upload_get_handler",
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "summary": "Get file from temporary storage",
        "tags": [
          "kamu"
        ]
      },
      "post": {
        "operationId": "platform_file_upload_post_handler",
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "items": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                },
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadContext"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "summary": "Upload file to temporary storage",
        "tags": [
          "kamu"
        ]
      }
    },
...
  },
...
}